Worked on some more abstraction of stuff into the Constants.h file for more control. All worked fine. It actually looks really good too.
Keep in mind when playing with this that I want the colour stuff left in for a good reason. I also added a toggle for backface culling because if you spawn underneath the ground surface it would be good if you could see it from below. Anyway more to come.